requirements.txt
numpy
tensorboard
tqdm~=4.43.0
ConfigArgParse
matplotlib~=3.7.5
plotly
shapely
scipy
setuptools~=70.2.0
scikit-learn
pillow
pyyaml~=6.0.2
pandas
joblib
sentence_transformers
wandb
trimesh
mesh_to_sdf
seaborn
tabulate
einops~=0.6.1
torchdiffeq
torch-summary
pygame
gitpython
scikit-video
opencv-python~=4.10.0.84
moviepy
hydra~=2.5
urdfpy~=0.0.22
gdown
einops

#torch~=2.0.0+cu118
networkx~=2.2
cholespy~=2.0.0
mmd~=0.1